+++ jhotdraw7/src/main/java/org/jhotdraw/io/StreamPosTokenizer.java
@@ -1,12 +1,9 @@
 /*
  * @(#)StreamPosTokenizer.java
  *
- * Copyright (c) 1999-2008 by the original authors of JHotDraw and all its
- * contributors. All rights reserved.
- *
+ * Copyright (c) 1999-2008 The authors and contributors of JHotDraw.
  * You may not use, copy or modify this file, except in compliance with the 
- * license agreement you entered into with the copyright holders. For details
- * see accompanying license terms.
+ * accompanying license terms.
  */
 package org.jhotdraw.io;
 
@@ -463,7 +460,7 @@ public class StreamPosTokenizer
         // rlw
         int data;
         if (unread.size() > 0) {
-            data = ((Integer) unread.lastElement()).intValue();
+            data = unread.lastElement().intValue();
             unread.removeElementAt(unread.size() - 1);
         } else {
             data = reader.read();
